phase1/requirements.txt
aiohttp==3.8.1
aiosignal==1.4.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==2.1.1
coverage==7.10.6
cryptography==46.0.1
exceptiongroup==1.3.0
filelock==3.19.1
flake8==7.3.0
frozenlist==1.7.0
fsspec==2025.9.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
multidict==6.6.4
mypy==1.16.1
mypy_extensions==1.1.0
numpy
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
propcache==0.3.2
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pytest==8.4.2
pytest-cov==7.0.0
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
tokenizers==0.22.0
tomli==2.2.1
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
